# openapi.yaml
openapi: 3.1.0 # Updated to OpenAPI 3.1.0 for access to newer features and improved JSON Schema support.
info:
  title: AI Teddy Bear API
  description: |
    Production-ready API for AI Teddy Bear - A child-safe, COPPA-compliant AI companion.
    
    ## Features
    - Child-safe AI interactions with content filtering
    - Full COPPA compliance with parental controls
    - Real-time voice interactions
    - Comprehensive safety monitoring
    - Educational content generation
    
    ## Security
    - JWT authentication with refresh tokens
    - Rate limiting on all endpoints
    - Data encryption for PII
    - Audit logging for compliance
  version: 1.0.0
  contact:
    name: AI Teddy Bear Support
    email: support@aiteddy.com # Consider externalizing contact information for dynamic updates in an enterprise setting.
  license:
    name: Proprietary
    url: https://aiteddybear.com/license
servers:
  - url: https://api.aiteddybear.com/v1
    description: Production server
  - url: https://staging-api.aiteddybear.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /children:
    get:
      tags:
        - Children
      summary: List all children for parent
      operationId: listChildren
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of children
          content:
            application/json:
              schema:
                type: object
                properties:
                  children:
                    type: array
                    items:
                      $ref: '#/components/schemas/Child'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Children
      summary: Create child profile
      operationId: createChild
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChildRequest'
      responses:
        '201':
          description: Child profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: COPPA violation or missing consent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /children/{childId}:
    get:
      tags:
        - Children
      summary: Get child profile
      operationId: getChild
      parameters:
        - $ref: '#/components/parameters/ChildIdParam'
      responses:
        '200':
          description: Child profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Children
      summary: Update child profile
      operationId: updateChild
      parameters:
        - $ref: '#/components/parameters/ChildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChildRequest'
      responses:
        '200':
          description: Child profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Children
      summary: Delete child profile
      operationId: deleteChild
      parameters:
        - $ref: '#/components/parameters/ChildIdParam'
      responses:
        '204':
          description: Child profile deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    ChildIdParam:
      name: childId
      in: path
      required: true
      description: Child profile ID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 50
        age:
          type: integer
          minimum: 3 # Age range aligned with COPPA compliance and target user group.
          maximum: 13 # Age range aligned with COPPA compliance and target user group.
        interests:
          type: array
          items:
            type: string
          maxItems: 10
        language:
          type: string
          enum: [en, es, fr, de, it, pt, zh, ja, ko]
          default: en
        personality_traits:
          type: array
          items:
            type: string
          maxItems: 5
        voice_enabled:
          type: boolean
          default: true
        safety_level:
          type: string
          enum: [strict, moderate, relaxed]
          default: strict
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_interaction:
          type: string
          format: date-time
          nullable: true
    CreateChildRequest:
      type: object
      required:
        - name
        - age
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        age:
          type: integer
          minimum: 3
          maximum: 13
        interests:
          type: array
          items:
            type: string
          maxItems: 10
        language:
          type: string
          enum: [en, es, fr, de, it, pt, zh, ja, ko]
          default: en
        personality_traits:
          type: array
          items:
            type: string
          maxItems: 5 # Enforces consistency with Child schema's maxItems for personality traits.
        consent_id:
          type: string
          format: uuid
          description: COPPA consent record ID
    UpdateChildRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        interests:
          type: array
          items:
            type: string
          maxItems: 10
        language:
          type: string
          enum: [en, es, fr, de, it, pt, zh, ja, ko]
        personality_traits:
          type: array
          items:
            type: string
          maxItems: 5 # Enforces consistency with Child schema's maxItems for personality traits.
        voice_enabled:
          type: boolean
        safety_level:
          type: string
          enum: [strict, moderate, relaxed]
...
    Error:
      type: object
      required:
        - error_id
        - message
      properties:
        error_id:
          type: string
          format: uuid
          description: Unique error ID for tracking
        message:
          type: string
          description: Human-readable error message
        detail:
          type: string
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: Request path that caused the error
  responses:
    BadRequest:
      description: Bad request. The request was malformed or invalid. Review the request body and parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized. Authentication credentials were missing, invalid, or expired. Provide a valid JWT token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found. The requested resource could not be found. Verify the resource ID or URL path.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...